Introduce client::Background (#46)
* Introduce client::Background

A public named type for the background task future, needed to
make the signature of the Service trait implementation of Connect
usable with generic executor types. Replaces ConnectExecutor.

* Restore ConnectExecutor

Respecified and blanket implemented in terms of client::Background.
Make it a public type as well.

* Addressed minor review comments

As per discussion in tower-rs/tower-hyper#46

* Internalize LiftBody in ConnectExecutor

ConnectExecutor is a utility trait to bind generic type parameters
as supplied by the API user and expose their bounds required
for an executor. Relegating part of it to LiftBody does not make sense.

* Internalized LiftBody in client::Background